Move Simulator to its own file
Additiona minor code cleanup.
